python_grpc/pyproject.toml
[tool.poetry]
name = "python_grpc"
version = "0.1.0"
description = ""
authors = ["Stefan Dinescu <aaetherengine@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.30.0"
grpcio-tools = "^1.30.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
